requirements.txt
streamlit==1.41.1
ultralytics==8.2.103
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4

streamlit-webrtc==0.47.7
av==14.1.0